Removed unwanted conversions for non-depth properties [devel]
Summary:
The internalValue vs userValue functions were introduced 12 years ago in
the UnitOfMeasure class, to make it such that the Feet UoM is the 'root' depth/distance
unit in depth projects which are in feet.
However the implementation was not limited to this case, primarily because it was based on
a single or double letter matching of the symbol. Thus switched it to a longer matching based on
the unit of measurement names.
With only this fix the implementation would still allow Feet based units for other data types
to be marked as internal, for instance Slowness, Impedance, Volume standard types.

I do not think this was intended, hence the first condition adds an additional constraint, to allow
only the Depth/Distance and Velocity types to remain in Feet or Feet/Second as internal units respectively.

Test Plan:
a conversion between seconds and milliseconds is now working correctly also
in depth projects in feet.

Reviewers: raman

Reviewed By: raman

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D15464