Fixed reading of layer properties
Summary:
Applying the UoM policy when reading the PropertyRef from the repository,
in PropertyRef::usePar
Also includes:
1- Fixes in the mnemonics repository: mismatch between the amplitudes and units for Compressibility,
and smaller value/unit changes
2- Removing redundant data in data/Properties (identical to the mnemonics, mostly the ranges),
added a symbol for all DefaultValue entries.
3- Fix for conversion of permeability between Dardy and MilliDarcy, by setting Darcy as the internal unit
for that type (Darcy is not SI).
4- The PropertyRef::uom_ was not copied from its mnemonic on construction of the object
5- Dissallowed having both a default value and a fixed definition.

[Jira: OD-4377]

Test Plan:
tested using F3_Demo (no .properties file in the project), and a client project with
heavily modified .properties file in the project. All values displayed in the layer properties
manager are in the expected unit of measurement.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D15469