Fix for crash/freeze while reading a large number of faults
Summary:
This is an extension of the fix in D15465. It fixes OD-4376.
Making sure the executors (reader/writer) are not sitting on open streams while idle. The fix is the main branch is going to be neater without the need for close/reOpen of streams.
Also fixed memory leaks for FaultSet objects.

Test Plan: Checked the number of open streams while reading/writing a large number of faults/horizons. Also tested HorizonData I/O. Checked the EMManager inventory after multiple epsodes of reading. There are no residual objects left.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D15470