Added Math::Formula::hasFixedUnits [devel]
Summary:
The formulas read in PropertyRefSet, and forwarded into
a PropertySet, were disconnected from the RockPhysics repository.
Thus added RockPhysics::FormulaSet::getMatching, to match a user-defined
expression, either typed by the user or read from the Properties file,
to an equation from the RockPhysics library. If a match is found the
RockPhysics formula properties are forwarded to the Math::Formula, which
provides the mnemonic and units for all inputs and the form output.
The formula will thus be able to return true to hasFixedUnits(), that
we can later reuse to hide the corresponding field from the GUI.

Test Plan:
Build an extensive test program function based on the formulas
most often used in Rock Physics. Almost complete, will also add the drawing
of some value to test the actual math calculation.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D15481