Added error message with incorrect layer properties selection
A basic check should be done before leaving uiSelectPropRefsGrp:
A valid input must be found for all properties which are based on
a fixed definition.

For instance it should not be allowed to select the compute MuRho
property without a Shear Impedance property in the set of selected
properties, if and only if MuRho is computed from SI.

Also corrected some errors in the error messages.