Added doPrepare and doFinish to Executor [devel]
Summary:
Added doPrepare for initialization and doFinish for cleanup. Made use of the new functions in a few classes. In the process, fixed the infinite loop in HorizonData import.
TODO: More classes need to be visited to look for doPrepare and doFinish candidates.

Test Plan: Tested HorizonData import, Faults to FaultSet workflow. All is well.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D15492