Better handling of non-DataRoot survey containers
Summary: While selecting a Free survey (when no licenses available), if the new data root did not have the mandatory omf file, there was an error message without a solution. Now the user is prompted to make the selected folder a data root. The dgb part of the fix is in D15498.

Test Plan: Cleared license, put a free project in a brand new folder, started OD, selected the new data root. A message pops up asking if I want to make it a data root. After OK and survey selection, OD starts up normally.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D15497