Improved the guessing of mnemonics [devel]
Summary:
Trying to stabilize the guessing of mnemonics from dataset names,
usually either log names or PDF axes names. The existing search methods were most often
stopping on the first match starting with the search name.
However lot's of mnemonic aliases are very short, leading the wrong returned matches.
Now the match returns a number that rates its match, thus that eventually the best
can be picked up.
Cleaned up the "GuessXX" functions, to keep only one per set type (set of Mnemonic or
set or PropertyRef objects): The search is typically done based on a string, but that
string can be replaced or complemented by the UnitOfMeasure or Mnemonic::Type object.
These two objects are used to get a relevant selection of possible outputs, limiting
the search scope and possible mismatch of the output. More search strings can be passed
in a set.
Updated the LAS import accordingly: The log curve are always the primary search key,
because they are mnemonics, and the curve description is only a hint. Since the LAS also
provides a unit of measure (most often), we now use all available information to set the
mnemonic when import a log from a LAS file.

Test Plan:
PDF axes are now correctly mapped on client project. Mnemonics shown in  the
well manager seem to be correct in a number of old project, meaning that they have been
assigned correctly during read from the database.
Started adding more tests to test_mnemonics, will be continued. Current tests pass.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D15505