Cleanup of the math formula GUI [devel]
Summary:
addressing various issues reported from clients:

- The 'Input unit' fields were misleading, and the user could not change their content. Removed them.
- The 'Formula requires' fields were also read-only for equations retrieved from the Rock Physics manager
or storage. They are now hidden (insensitive in debug mode to allow QC). However to still allow
advanced user to display, QC, change them, a toggle was added, that will only appear when these fields are
hidden.

- The same applies to the field 'Formula result is'.
- The form output mnemonic selection was form too large and not handy. Thus added a Mnemonic::StdType field
that acts as a filter, for a more efficient selection of the output mnemonic.

- Reintroduced the option to select no unit of measure for either an input or the formula output. This remains
a shortcut, and equivalent to setting the formula unit to the internal unit of measure for that type.

- Added a matching of the formula text input by the user to the equations in the rock physics library. The
match will have to be quite precise, as only a minor cleanup in done to handle white spaces and the case.

Same as in the main branch, except that the Math::Formula class does not get the 'allChanged' notifier.

Test Plan:
same as D15543
+ No warning related to hiddenparam usage

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D15544