Added environment variables to control MMP listening server
Summary:
On a client workstation (Linux, gcc5 based), the listening server
started by od_SeisMMBatch is strangely only listening to IPv6 connections.
This is stated by the ss -tulpn command, which reports [::]:37500 with our
default port for MMP, whereas on Linux it should be in the format *:37500,
which means listening to both IPv4 and IPv6 connections.
Hence added the following environment variable that can be set to control
which type of interface the server should be listening too:
DTECT_MMP_IPv6: listens to only IPv6
DTECT_MMP_IPv4: listens to only IPv4
Both default to false in their absence.

Test Plan:
On the client workstation, set DTECT_MMP_IPv4 to Yes does create
a server that listens to all interfaces (sic): It is reported as "*:37500".
Whilst this is still odd, it does allow incoming IPv4 connections, giving a
chance for the processing nodes to connect with the client app od_SeisMMBatch

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D15545