Fix for an extra undef line read during ascii import
Summary:
The problem manifests itself as an extra line with 1e30 values when you convert XY/LL positions. But it was a general problem, not detected during other imports as the undef vallues were discarded.
Also removed the unnecessary initialization of msg_ as that was making all the 'msg.isEmpty()' checks redundant.

Test Plan: Converted LL to XY. No extra lines seen at the end. Exported a PointSet, imported back and confirmed in the debugger that there were no undef values coming at the end, and also that no legit positions were missed.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D15550