Restored return behavior of IOPar::get(key,string) functions
Summary:
In case of 'Empty string not allowed, a valstr.isEmpty() check works well. But when empty strings are allowed, isPresent(key) is needed to distringuish between a missing entry and an empty entry.
Also removed unused macros.

Test Plan: Tested Destriping process successfully, which had triggered this change in the first place.

Reviewers: nanne

Reviewed By: nanne

Subscribers: prajjaval

Differential Revision: https://phabricator.dgbes.com:8443/D15564