Enabled setting of project system from SEG-Y SIP [devel]
Summary:
One should be able to set the projection system when creating a
project from a SEG-Y file. The GUI had the corresponding interface but it was
hidden in survey creation mode, and uiSEGYSurvInfoProvider did not reimplement
getCoordSystemPars.
This will allow the user to set the projection by looking at the content of the textual header.
In the few cases when the SEG-Y file was produced by OpendTect, the information is
automatically retrieved from the header.

[Jira: OD-4431]
[Jira: OD-4432]
[Jira: OD-4433]
[Jira: OD-4434]
[Jira: OD-4435]

Test Plan: Can retrieve and set the information from a SEG-Y file during survey setup

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D15573