Preliminary python bindings using ctypes
Summary:
Prelim look at making a set of Python bindings not tied to a particular version of Python
using the ctypes module. The binding shared object is built like a normal OpendTect plugin and
there is a companion python module (python/odbind.py) which provides the corresponding Python code
to access methods in the shared object library. The python folder also contains a Jupyter notebook
that demonstrates the use of the binding. Only basic acess to Surveys and 2D horizons at the moment.
It should work if you open the notebook from a JupyterLab session stated from OpendTect.

Having to provide a C level library interface is a pain but an advantage would be it opens up
relatively easy binding to other languages that provide a foreign function interface like
Julia and even Octave/Matlab.

Fix lint issues

Test Plan: Tested on Linux

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D15579