Fix for lone seed on shift-click while picking Fault Sticks
Summary: The case of Shift-click is already handled below. There is no need to set the flag makenewstick_ to true, as that is meant for signalling the 'next' event to start a new stick.

Test Plan: Picked a few sticks using different cobinations of click, ctrl-click, double-click, shift-click etc. Everything works as expected.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D15598