Fixed incorrect check for C++ ABI in init_dtect script
The string checked within the libstd library should be based on CXXABI,
neither GLIBC not GLIBCXX, as they may give incomplete results.

Now setting the search to be based on CXXABI_1.3.12, which is the ABI version which
matches the ABI version of the compiler used to build this branch (gcc 9.5.0)

Also fixed a syntax error on Ubuntu, and made the script compatible with the
possible lack of installed "strings" system command.

Now OpendTect starts on RHEL 7.X, and 8.X, even if the installed C++
standard library is less than 9.

The required version should be updated using the following table:
https://gcc.gnu.org/onlinedocs/libstdc++/manual/abi.html#abi.versioning