Fix for Session Restore with 2D Viewers
Summary:
While restoring a Scene, the childids where sorted, but the IOPars were not. As a result, the ids were assigned to different objects than in the saved session. Since the 2D viewers are linked to the 3D visIDs, they were not able to restore.
Fixes OD-4462

Test Plan: Created a scene with Inlines, Crosslines, PointSets, Horizons and 2D viewers for both Inlines and Crosslines. I was able to save and restore it completely.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D15623