Executing 'strip' command after extracting symbols on Linux
Summary: We are using the '-g' flag on Linux even in Release mode in order to have the debug symbols in the binaries that are extracted using dump_syms. After the execution of dump_syms, we need to run 'strip' on the binaries to strip them of the debug symbols. We were earlier stripping using Linker flags which caused the extracted symbols to not have the source code references.

Test Plan: Tested in od7.0 by running a release build. The extracted symbols have all the required info including the source code references, whereas the final binaries are stripped successfully.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D15654