Added cmake setup for safety package and odbind [devel]
Adding the safety package and its marshmallow dependency to the setup
within a developments environments.
Adding the installation of both python packages
Adding the copy of these two packages, and odbind.py to the base package

Verified that the version checked out (2.4.0 for safety) is the one used
by the safety dialog within OpendTect, even if another safety package is
present within the conda environment.