Attempt to fix 'Show in folder' functionality on Linux
Summary: 'Show in folder' does not work for me on OpenSuse in release as the file manager program dolphin cannot find the required dependencies in an OD runtime environment. Running the file manager in isolated mode should solve this issue.

Test Plan: Works as usual in my development env, need to test in a release.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D15671