Fix for multi-component attribute output with custom sampling
Summary: This bug affected all the multi-component attribute outputs when the output sampling was different from the input sampling. The reSize call was made on individual component. This led to wrong output for the second compnent onwards.

Test Plan: Computed SpecDecomp, HorizonCube-To-SteerinCube with samplings different from the survey sampling. The values of all the components are correct now.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D15691