Fixed import-copy of d2tmodels in projects with large SRD
Summary:
The time-depth models in projects when the wells datum (KB) is
below SRD (typically onshore projects with large SRD) are such that the TWT
value of all points in the model is function of both SRD and the well KB.

As a result, in order to correctly import such d2tmodels (from ascii or accross wells),
one must assume that the incoming model is also coming from a well where SRD > KB.
Such case is now taken into account when mapping the incoming model on the track
of the target well.

Also found and fixed accuracy issues: The table now contains a truncated version of the
d2tmodel, it is thus not accurate enough to recreate the model from the screen display.
Hence adding a caching of the full accuracy data.

[Jira: OD-4553]

Test Plan:
Using a project provided by a consultant. After import/copy the d2tmodel is
now in line with its source.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D15719