Added support for user-defined mnemonics
Summary:
We need some way for the user to increase the list of mnemonics,
but with a controlled mechanism, not using the multi repository mechanism that we
have for the properties.
Here I propose to setup a project-specific list with a simple GUI that allows the
selection of one 'template' mnemonic from the software list, assign any name to it,
save it in the survey .defs file. It gets retrieved on the first call to the global MNC(),
cleared on survey change, reloaded in the new project.
For now the option is added to the Basic well manager. I will check with the user
where else it needs to become available.

[Jira: OD-4494]

Test Plan:
Tested and showed the GUI to the reporter. Load, save, revert actions are
working correctly.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D15776