Fixed distributed processing error
Summary:
The path conversion implementation in the HostData class was
wrong, leading to many values in the IOPar of each job to start with a
string contains twice the data root, the fromprefix and the toprefix as well.
I am replacing the implementation with higher level functions from BufferString.
Can you confirm the implementation?

[Jira: OD-4608]

Test Plan: Tested linux to linux processing, working fine to localhost and remote node.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D15810