Restricted support for Anchor based Coordinate system
Summary: The 'Anchor Based XY' will not appear in the selection list unless the current CRS is of this type.

Test Plan: Tested the CRS tab in Survey info editor for various surveys. The 'Anchor based XY' entry is present only when the current survey has this coordinate system set.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D15814