Fixed distributed processing error
Summary:
The path conversion implementation in the HostData class was
wrong, leading to many values in the IOPar of each job to start with a
string contains twice the data root, the fromprefix and the toprefix as well.
I am replacing the implementation with higher level functions from BufferString.

[Jira: OD-4608]

Test plan:
Tested linux to linux processing, working fine to localhost and remote node.