Prevent automatic loading of all horizon data
Summary:
Currently loading the geometry of a 3D horizon to display the Z values also loads all
associated horizon data regardless of what the user selects. For horizons with a lot of horizon data
there is a significant penalty in loading time and memory usage. Proposed change ensures users horizon data
selection is not overwritten and actually used.

Test Plan:
Tested on Linux:
- display of 3D horizons and horizon data still works

Reviewers: nanne, raman

Reviewed By: raman

Subscribers: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D15829