Fixed test_callback program [devel]
In release mode, randomly, the program did not stop. The timer callback function
NotifiedClass::closeTesterCB was never hit, although NotifiedClass::timerHit was
correctly processed.
Very likely the presence of two ApplicationData objects was distubing the events
processing.
Removed the local ApplicationData object in the InMainThreadTester::test static
function. It should not be placed here. Also the previous logic was not in line
with the documentation of this class. Threads::WorkManager::addWork requires an
explicit argument to have to work done on a different thread.

Ran the test_program many times in debug and release mode on Linux, seems to be
working correctly now.