Fixed saving models with updates in SynthRock
The function uiLayerSequenceGenDesc::editedDesc() was not virtual,
hence uiLayerSequenceGenDesc::currentDesc() was always using the base implementation,
triggering the issue.
Very likely the Go button must be triggering a full copy of the edited desc to the
base desc, which would explain the workaround of pressing the Go button before saving.

[Jira: OD-4631]