Getting a global list of all the well-markers sorted in depth
Summary: Added a function to merge ordered lists of strings. In this case, it is being used for well-markers, which are always depth-sorted for a particula well. For now, it is being used in the 'Create attribute log' window. Next I am going to use it in WellTable.

Test Plan: Tested the list of markers in the Create attribute log window. Also printed and checked the relationmap array for consistency.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D15835