ErrMsg should print to std::cerr in standard apps
Summary:
The python link parses the (std) output of od_DBMan for a number of
queries, like getting the filename of a MultiID implementation. The normal
output is a string that can be parsed into a JSON object.
However this breaks down if the standard output is corrupted by the presence of other messages.
I added a filter inside odpy to isolate the json string returned in the standard output, but
I consider this patch as the core fix: The ErrMsg function should only using the log stream
if the global log stream has been set to a file. This is the case only for od_main, any other
application used std::cout, and for such apps the output of ErrMsg should be directed to
std::err, neither to std::cout nor to the file behind od_ostream::logStream().

[Jira: OD-4651]

Test Plan: Inserted a dummy message send using ErrMsg at application startup. This simulates what happens when the loading of a plugin fails (not fatal). Then ran the command 'od_DBMan --status 2> error.txt 1> out.txt', and checked that the message is written to error.txt, not out.txt.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D15868