Added separate OD::Platform::Type for macOS ARM and Intel
Summary: Apart from making the code aware of the two different macOS architectures, the goal here is also to support separate binary packages for the two platforms in the Installer. Since the Installer still needs to support v6.6 packages, we need to keep the type Mac as it is, and just add two new entries MacARM and MacIntel for v7.0.

Test Plan: compiles on linux. Tested the mac specific preprocessor macros using simple programs on the two macs.

Reviewers: nanne, arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D15872