Running conda list in the background
Summary:
The event loop locking at startup is coming from the command 'conda list' not being executed in the background. Using the thread work manager to run this command in the background, the first time the PythonAccess class is being used.
If the application is closed during the execution in the background of this command, ApplicationData::hasInstance allows avoiding to parse the output, or sending a signal.
There is no scope in skipping PythonAccess::updateModuleInfo, as a python linking without information on the available modules is of little use.

Test Plan: tested from the plugins. The functions attached to PythonAccess::envChange are no longer hit when the application is closing during the processing of this command (takes about 10sec max on Windows).

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D15926