Fixed incorrect display scaling while changing synthetic data wavelets
Switching the active wavelet of a synthetic dataset means that the amplitude range
of the synthetic dataset may change enormously. Hence the cached mapper range may
only be kept if the previous and next synthetic datasets have about the same scale,
which is checked now by comparing the RMS of their wavelets
Working on their wavelets makes it very fast, and insensitive to the reflectivities
that are present in the pseudowells.