Fixed initial survey setup #2
The modal condition for uiTaskRunner was incorrect, it should be modal in
the absence of a parent. Furthermore the uiTaskRunner object may be setting
a different parent than the provided first argument during construction.
Hence checking if the effective parent is null.
This is actually extremely important, as the uiTaskRunner will crash in the absence
of a parent when non modal.

[Jira: OD-4677]