Fixed crash when importing 2D data
Summary:
Only happens when importing to CBVS but give an output attribute name that
already exists as SEGYDirect.

It looks like this class is never used.
tr_ in SEGYDirect2DLinePutter is never created and wasn't initialized.

Test Plan:
No crash anymore. But we need a better mechanism to prevent the user to give
a name that already exists as another type.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D15949