Fixed several shift issues in SynthRock because of SRD
Summary:
Several items did not correctly take into account the seismic
reference datum in the SynthRock viewer:
1- The computation of the time of markers
2- The lookup of the sequence layer corresponding to a certain TWT value
3- The synchronized zoom between time and depth viewers.
On this last point, the MFVCViewManager should not be directly using SI()
in its operations. It is up to the caller to provide the required scaling
information, i.e. uiStratSynthDisp in this case.

[Jira: OD-4701]

Test Plan:
Using a project with a large SRD (750m), that shows obvious shifts.
That start view is correct, rubberband zoom is synchronized, with an without
flattening, from either time or depth domain.
The markers in a time are placed where there should be.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D15958