Updated implementation of System::localAddress
Summary:
System::localAddress was returning the first usable IPv4 address,
but this address may be the address used by a VPN adapter, and MMP processing nodes cannot respond back to the client using its VPN address, it needs to get either the Ethernet or the Wifi network address.
Updated the implementation of System::localAddress: Instead of relying on the DNS to return the appropriate IP address for the local host, making a loop over all local interfaces, and establishing a preference list.

Furthermore, if the client machine is listed in the setup for distributed processing, using the IP address of that setup as the preferred local address.

Should be the last step missing for:
[Jira: OD-864]
[Jira: OD-2961]
[Jira: OD-3964]
[Jira: OD-4721]

Test Plan:
Now getting the Wifi IPv4 address as the computer address, while the VPN connection is open.
Tested distributed processing with:
Ethernet connection up, with or without OpenVPN connection, or Wifi connection up, with or without OpenVPN connection
Local host missing from the configuration file, or present by either host name or IP address. When present, the IP address was the Ethernet LAN IP address, which is invalid when the cable is not connected.
Airplane mode and no configuration file

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D15982