Fixed distributed processing testing error
Removed the incorrect error 'default data root not present', for nodes that were
configured with a common unit data root only.
Reworked several error messages.
During testing, uiBatchHostsDlg will set the data root on the processing nodes (od_remoteservice)
if different from the configuration, as this is still a requirement for the processing to work.
Therefore, added the tools that will update the od_RemoteServiceMgr GUI when the survey data
root of od_remoteservice is changed.

[Jira: OD-4727]