Fixed multiple Horizon Data writing
Summary:
This was originally reported by Glen. After we implemented a delayed initialization of streams for reader/writer classes, the HOV file was not created until the corresponding executor started executing. This resulted in the getFreeFilename() function returning the same file name for all the attributes. Now delaying the getFreeFilename() function call solves this problem, as by that time the previous files are already written to disk and the getFreeFilename function returns a new file name for each Horizon data.
Also renamed Attributes in import menu to Horizon Data.

Test Plan: Imported 4 Horizon Data for a horizon and displayed them in the scene. Each Horizon Data is written to separate files and the values also reflect the ones from the imported file.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D16003