Passing Storereqs as a reference now, was passed by value before
Summary:
Storreqs in store functiono of MultiWellReader was passed by value and making a copy. Secondly, the value passed was not storereq itself but that of pointer. This meant that storereqs values changed which was causing all the files for a well to be rewritten where only the marker file should have.

Jira Issue: OD-4730

Porting marker fixes from main

Test Plan: tested in well table

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: wayne, lloyd

Differential Revision: https://phabricator.dgbes.com:8443/D16012