Fixed issues in launching of od_remoteservice
Summary

launching od_remoteservice when no data root is set in the settings did not work
launching od_remoteservice using the key dataroot did not work when no active survey is set in the settings
launching od_remoteservice when there is already another instance running did not give an error
launching od_RemoteServiceMgr with an already running od_remoteservice connected to
a survey data root different from the user settings did work, but the GUI was
showing the survey data root from the settings and not from the server app od_remoteservice

Added od_RemoteServiceMgr in the menu Utilities -> Tools -> Batch Programs,
made it the default.
Moved od_remoteservice just below it. Ensured it no longer starts an empty progress viewer
Added a new flag for machine commands from which one only wants to display the standard output.
Removed the code that uses the processing progress viewer to display it, a custom uiMainWin
should be used instead.

[Jira: OD-4752]
[Jira: OD-4753]

Test Plan
To solve it, because neither od_remoteservice not od_RemoteServiceMgr need a valid survey,
I made both apps to not require a valid data directory (third argument of SetProgramArgs true - false).
Not calling IOMan::setDataSource anymore, as this can only be done when having also a valid survey,
which is not necessarily the case.
od_remoteservice detects an already used port, and shows and appropriate error message
od_RemoteServiceMgr reads the mmplient data root and set the GUI accordingly if OK.