Fix for v7 horizons not displaying seeds in v6
Summary: We got rid of section IDs in v7. But v6 needs the section ids in order to read the seeds. Therefore, putting the default section IDs in the IOPar fixes the problem.

Test Plan: Checked two-way compatibility between v6.6 and v7.0. Seeds picked in either version are displayed correctly in the other version.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D16049