Fixed cmake error when trying to find the 3rd party packages [devel]
The OD_FIND_XXX macros are responsible of finding the 3rd party packages using find_package
and defining their corresponding targets in the main scope of the project,
however they should not fail and should always continue.
The actual check needs to be placed in the OD_SETUP_XXX macros, which are only called in
the modules linking or using the externals. And even there one should not issue a FATAL_ERROR,
but a SEND_ERROR, to collect all errors from the generation at once, at populate the cache variables
alltogether.

Also forwarding now the SQLITE_DIR variable to FindOpendTect, such that it is knows by external projects
that link against the OpendTect basic modules (like all closed source projects).
Finally, removing the option to build proj as an external, it would require too much maintenance to keep it.