Fixed horizon picking on random lines
Summary:
Picking 1 seed resulted in a flat horizon, instead of a tracked line.
In one of the checks the datapack was casted to a RegularSeisDataPack, but
random lines use a RandomSeisDataPack

OD-4761

Test Plan:
Added random line, picked some seeds and pressed the auto-track button.
This now works as expected.

Reviewers: wayne

Reviewed By: wayne

Subscribers: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D16073