Reworked the cmake scripts for creating release packages [devel]
Summary:
changes required to make the packaging work after the changes
to support the multi config generators, and large cleanup of the
corresponding code.
This should also fix the packaging of the Qt dependencies on macOS, and
enable shipping the QtWebEngine on this platform too.

Test Plan:
packages are created successfully on Linux, will need
testing on Windows and macOS.

Reviewers: raman

Reviewed By: raman

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D16084