Added loading of all SIP implementations when loading uiIo
Summary:
The cmake macro OD_INSTALL_SIP can be used (after OD_INIT_MODULE) to
create a file specifying that a module or plugin has one or several uiSurvInfoProvider
implementations to be added to the factory.
If that is the case, the module or plugin should define and implement a void function
named ${OD_MODULE_NAME}InitSIP(). This function will be searched by uiIo during its
initialization, and called if found.
This technique should be applied by all modules higher than uiIo, and all plugins that need it.

[Jira: OD-4797]

Test Plan:
Tested creation of a project in an empty data root from a SEGY file. The project
is correctly created, but the steps thereafter have issues specific to the main branc,
probably unrelated to this issue.

Reviewers: nanne

Reviewed By: nanne

Subscribers: prajjaval, raman

Differential Revision: https://phabricator.dgbes.com:8443/D16111