Apply to all now works in Default log dialog
Summary:
"Apply to all wells" button wasn't working in the editing Default Logs for Mnemonic dialog. This was happening because uiComboBox notifier was being triggered first and that meant that currentCell function of the table was not up to date.

I am now dynamically casting the Callbacker to identify which ComboBox is triggering the notifier. This fixes the issue.

Test Plan: tested in "Set/Edit Default Logs for Mnemonic" dialog in Well Table.

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: wayne, lloyd

Differential Revision: https://phabricator.dgbes.com:8443/D16119