Fixed order of menu items in Survey Import Seismic
Summary:
This menu contains sub-menus and (direct) actions.
The order was not set at all between the various items added, leading to
inconsistencies when the loading order of the modules and plugins changes.
First there was no API for adding a menu after a uiAction, thus added it.
Secondly I am changing the order of the items:
Always SEG-Y first, then CBVS, then the less frequency used items (Simple file and GPR).

[Jira: OD-4828]

Test Plan:
The order is this menu is now always the same, on boot and after
changing the survey.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D16125