displaying color from .pck file
Summary:
Earlier the in the absence of disp file, default color was been assigned to picks on display. But the color was also stored in the .pck file. This color as been read when the data was read but was overwritten because of the absence of display file. Now proper check are placed.
Also upon double click the and changing the color, the color was not updated in the treeviewitem. OD-4806

Test Plan:
Color property was deleted from one file and kept in anpther file. No disp file was present in both cases. In first case where no color property was present in .pck file Red color was assigned and in case where Color property was present the same color was displayed
Color also updated in treeitem

Reviewers: arnaud

Reviewed By: arnaud

Subscribers: raman, lloyd, wayne

Differential Revision: https://phabricator.dgbes.com:8443/D16136