Fixed use of saved formats when importing horizons
Summary:
This is a mess! Apparently, the strings used in our GUI need to be identical to the strings used in the FileFormats file!

Had to add the 'Position' label again. Moved the formfld_ for better layout.

Jira: OD-4837

Test Plan: In import 3D horizons, selected several formats. Result looks OK now.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D16150