Fix ODBind memory allocation issue on Windows
Summary:
Fixes OD-4732.
Wasn't able to test that the Crossplot app is working within the OpendTect Qt webview window on Windows
because my Windows system is too underpowered to start the Bokeh server before the window is displayed.
However, was able to connect to the Bokeh server from a web browser and confirm the crossplot app is
working again.

Test Plan:
Tested on Linux and Windows (see comments). All python tests passing on Linux some minor fails on
Windows that I will investigate when time permits.

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D16197